requirements.txt
torch
torchvision
pyyaml
tqdm